Mux e2e tests
