--- e2e/test/MessageReceive.cs
@@ -1,71 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.Azure.Devices.Client;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.Azure.Devices.E2ETests
-{
-    public static class MessageReceive
-    {
-        private static TestLogging s_log = TestLogging.GetInstance();
-
-        public static Message ComposeC2DTestMessage(out string payload, out string messageId, out string p1Value)
-        {
-            payload = Guid.NewGuid().ToString();
-            messageId = Guid.NewGuid().ToString();
-            p1Value = Guid.NewGuid().ToString();
-
-            s_log.WriteLine($"{nameof(ComposeC2DTestMessage)}: payload='{payload}' messageId='{messageId}' p1Value='{p1Value}'");
-
-            return new Message(Encoding.UTF8.GetBytes(payload))
-            {
-                MessageId = messageId,
-                Properties = { ["property1"] = p1Value }
-            };
-        }
-
-        public static async Task VerifyReceivedC2DMessageAsync(Client.TransportType transport, DeviceClient dc, string payload, string p1Value)
-        {
-            var wait = true;
-
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
-            while (wait)
-            {
-                Client.Message receivedMessage = null;
-
-                receivedMessage = await dc.ReceiveAsync(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
-
-                if (receivedMessage != null)
-                {
-                    string messageData = Encoding.ASCII.GetString(receivedMessage.GetBytes());
-                    s_log.WriteLine($"{nameof(VerifyReceivedC2DMessageAsync)}: Received message: {receivedMessage}");
-
-                    Assert.AreEqual(payload, messageData);
-
-                    Assert.AreEqual(1, receivedMessage.Properties.Count);
-                    var prop = receivedMessage.Properties.Single();
-                    Assert.AreEqual("property1", prop.Key);
-                    Assert.AreEqual(p1Value, prop.Value);
-
-                    await dc.CompleteAsync(receivedMessage).ConfigureAwait(false);
-                    break;
-                }
-
-                if (sw.Elapsed.TotalMilliseconds > FaultInjection.RecoveryTimeMilliseconds)
-                {
-                    throw new TimeoutException("Test is running longer than expected.");
-                }
-            }
-
-            sw.Stop();
-        }
-    }
-}
