+++ e2e/test/MethodE2EMultiplexingOverAmqpTests.cs
@@ -13,20 +13,14 @@ namespace Microsoft.Azure.Devices.E2ETests
 {
     [TestClass]
     [TestCategory("IoTHub-E2E")]
-    public class MethodE2EMultiplexingTests : IDisposable
+    public class MethodE2EMultiplexingOverAmqpTests : IDisposable
     {
-        private readonly string DevicePrefix = $"E2E_{nameof(MethodE2EMultiplexingTests)}_";
-        private readonly int MuxWithoutPoolingDevicesCount = 2;
-        private readonly int MuxWithoutPoolingPoolSize = 1; // For enabling multiplexing without pooling, the pool size needs to be set to 1
-        private readonly int MuxWithPoolingDevicesCount = 4;
-        private readonly int MuxWithPoolingPoolSize = 2;
-
-        // TODO: #839 - Implement proxy and mux tests
-        private static string ProxyServerAddress = Configuration.IoTHub.ProxyServerAddress;
-        private static TestLogging _log = TestLogging.GetInstance();
+        private const string MethodName = "MethodE2EMultiplexingOverAmqpTests";
+        private readonly string DevicePrefix = $"E2E_{nameof(MethodE2EMultiplexingOverAmqpTests)}_";
         private readonly ConsoleEventListener _listener;
+        private static TestLogging _log = TestLogging.GetInstance();
 
-        public MethodE2EMultiplexingTests()
+        public MethodE2EMultiplexingOverAmqpTests()
         {
             _listener = TestConfig.StartEventListener();
         }
@@ -37,9 +31,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -49,9 +43,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -61,9 +55,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -73,9 +67,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -85,9 +79,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithPoolingPoolSize, 
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize, 
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -97,9 +91,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -109,9 +103,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -121,9 +115,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethod
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethod
                 ).ConfigureAwait(false);
         }
 
@@ -133,9 +127,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -145,9 +139,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -157,9 +151,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -169,9 +163,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.Device,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -181,9 +175,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -193,9 +187,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithoutPoolingPoolSize,
-                MuxWithoutPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithoutPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithoutPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -205,9 +199,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_Tcp_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -217,9 +211,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespondMuxedOverAmqp(
                 ConnectionStringAuthScope.IoTHub,
                 Client.TransportType.Amqp_WebSocket_Only,
-                MuxWithPoolingPoolSize,
-                MuxWithPoolingDevicesCount,
-                MethodOperation.SetDeviceReceiveMethodDefaultHandler
+                MultiplexingOverAmqp.MuxWithPoolingPoolSize,
+                MultiplexingOverAmqp.MuxWithPoolingDevicesCount,
+                MethodE2ETests.SetDeviceReceiveMethodDefaultHandler
                 ).ConfigureAwait(false);
         }
 
@@ -228,65 +222,40 @@ namespace Microsoft.Azure.Devices.E2ETests
             Client.TransportType transport,
             int poolSize,
             int devicesCount,
-            Func<DeviceClient, Task<Task>> setDeviceReceiveMethod
+            Func<DeviceClient, string, Task<Task>> setDeviceReceiveMethod
             )
         {
-            var transportSettings = new ITransportSettings[]
+            Func<DeviceClient, TestDevice, Task> initOperation = async (deviceClient, testDevice) =>
             {
-                new AmqpTransportSettings(transport)
-                {
-                    AmqpConnectionPoolSettings = new AmqpConnectionPoolSettings()
-                    {
-                        MaxPoolSize = unchecked((uint)poolSize),
-                        Pooling = true
-                    }
-                }
+                _log.WriteLine($"{nameof(MethodE2EMultiplexingOverAmqpTests)}: Setting method for device {testDevice.Id}");
+                Task methodReceivedTask = await setDeviceReceiveMethod(deviceClient, MethodName).ConfigureAwait(false);
             };
 
-            ICollection<DeviceClient> deviceClients = new List<DeviceClient>();
-            Dictionary<DeviceClient, int> deviceClientConnectionStatusChangeCount = new Dictionary<DeviceClient, int>();
-
-            try
+            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
             {
-                for (int i = 0; i < devicesCount; i++)
-                {
-                    ConnectionStatus? lastConnectionStatus = null;
-                    ConnectionStatusChangeReason? lastConnectionStatusChangeReason = null;
-                    int setConnectionStatusChangesHandlerCount = 0;
+                _log.WriteLine($"{nameof(MethodE2EMultiplexingOverAmqpTests)}: Preparing to receive method for device {testDevice.Id}");
+                await MethodE2ETests.ServiceSendMethodAndVerifyResponse(
+                    testDevice.Id, MethodName, MethodE2ETests.DeviceResponseJson, MethodE2ETests.ServiceRequestJson).ConfigureAwait(false);
+            };
 
-                    TestDevice testDevice = await TestDevice.GetTestDeviceAsync($"{DevicePrefix}_{i}_").ConfigureAwait(false);
-                    DeviceClient deviceClient = testDevice.CreateDeviceClient(transportSettings, authScope);
-                    deviceClients.Add(deviceClient);
-
-                    deviceClient.SetConnectionStatusChangesHandler((status, statusChangeReason) =>
-                    {
-                        setConnectionStatusChangesHandlerCount++;
-                        lastConnectionStatus = status;
-                        lastConnectionStatusChangeReason = statusChangeReason;
-                        _log.WriteLine($"{nameof(MethodE2EMultiplexingTests)}.{nameof(ConnectionStatusChangesHandler)}: status={status} statusChangeReason={statusChangeReason} count={setConnectionStatusChangesHandlerCount}");
-                        deviceClientConnectionStatusChangeCount[deviceClient] = setConnectionStatusChangesHandlerCount;
-                    });
-
-                    Task methodReceivedTask = await setDeviceReceiveMethod(deviceClient).ConfigureAwait(false);
-                    await Task.WhenAll(
-                        MethodOperation.ServiceSendMethodAndVerifyResponse(testDevice.Id),
-                        methodReceivedTask).ConfigureAwait(false);
-                }
-            }
-            finally
+            Func<IList<DeviceClient>, Task> cleanupOperation = async (deviceClients) =>
             {
-                // Close and dispose all of the device client instances here
                 foreach (DeviceClient deviceClient in deviceClients)
                 {
-                    await deviceClient.CloseAsync().ConfigureAwait(false);
-
-                    // The connection status change count should be 2: connect (open) and disabled (close)
-                    Assert.IsTrue(deviceClientConnectionStatusChangeCount[deviceClient] == 2, $"Connection status change count for deviceClient {TestLogging.GetHashCode(deviceClient)} is {deviceClientConnectionStatusChangeCount[deviceClient]}");
-
-                    _log.WriteLine($"{nameof(MethodE2EMultiplexingTests)}: Disposing deviceClient {TestLogging.GetHashCode(deviceClient)}");
                     deviceClient.Dispose();
                 }
-            }
+                await Task.FromResult<bool>(false).ConfigureAwait(false);
+            };
+
+            await MultiplexingOverAmqp.TestMultiplexingOperationAsync(
+                DevicePrefix,
+                transport,
+                poolSize,
+                devicesCount,
+                initOperation,
+                testOperation,
+                cleanupOperation
+                ).ConfigureAwait(false);
         }
 
         public void Dispose()
