--- e2e/test/MethodOperation.cs
@@ -1,123 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.Azure.Devices.Client;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.Azure.Devices.E2ETests
-{
-    public static class MethodOperation
-    {
-        private const string DeviceResponseJson = "{\"name\":\"e2e_test\"}";
-        private const string ServiceRequestJson = "{\"a\":123}";
-        private const string MethodName = "MethodE2ETest";
-        private static TestLogging s_log = TestLogging.GetInstance();
-
-        public static async Task ServiceSendMethodAndVerifyResponse(string deviceName)
-        {
-            using (ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
-            {
-                s_log.WriteLine($"{nameof(ServiceSendMethodAndVerifyResponse)}: Invoke method {MethodName}.");
-                CloudToDeviceMethodResult response =
-                    await serviceClient.InvokeDeviceMethodAsync(
-                        deviceName,
-                        new CloudToDeviceMethod(MethodName, TimeSpan.FromMinutes(5)).SetPayloadJson(ServiceRequestJson)).ConfigureAwait(false);
-
-                s_log.WriteLine($"{nameof(ServiceSendMethodAndVerifyResponse)}: Method status: {response.Status}.");
-                Assert.AreEqual(200, response.Status);
-                Assert.AreEqual(DeviceResponseJson, response.GetPayloadAsJson());
-
-                await serviceClient.CloseAsync().ConfigureAwait(false);
-            }
-        }
-
-        public static async Task<Task> SetDeviceReceiveMethod(DeviceClient deviceClient)
-        {
-            var methodCallReceived = new TaskCompletionSource<bool>();
-
-            await deviceClient.SetMethodHandlerAsync(MethodName,
-                (request, context) =>
-                {
-                    s_log.WriteLine($"{nameof(SetDeviceReceiveMethod)}: DeviceClient method: {request.Name} {request.ResponseTimeout}.");
-
-                    try
-                    {
-                        Assert.AreEqual(MethodName, request.Name);
-                        Assert.AreEqual(ServiceRequestJson, request.DataAsJson);
-
-                        methodCallReceived.SetResult(true);
-                    }
-                    catch (Exception ex)
-                    {
-                        methodCallReceived.SetException(ex);
-                    }
-
-                    return Task.FromResult(new MethodResponse(Encoding.UTF8.GetBytes(DeviceResponseJson), 200));
-                },
-                null).ConfigureAwait(false);
-
-            // Return the task that tells us we have received the callback.
-            return methodCallReceived.Task;
-        }
-
-        public static async Task<Task> SetDeviceReceiveMethodDefaultHandler(DeviceClient deviceClient)
-        {
-            var methodCallReceived = new TaskCompletionSource<bool>();
-
-            await deviceClient.SetMethodDefaultHandlerAsync(
-                (request, context) =>
-                {
-                    s_log.WriteLine($"{nameof(SetDeviceReceiveMethodDefaultHandler)}: DeviceClient method: {request.Name} {request.ResponseTimeout}.");
-
-                    try
-                    {
-                        Assert.AreEqual(MethodName, request.Name);
-                        Assert.AreEqual(ServiceRequestJson, request.DataAsJson);
-
-                        methodCallReceived.SetResult(true);
-                    }
-                    catch (Exception ex)
-                    {
-                        methodCallReceived.SetException(ex);
-                    }
-
-                    return Task.FromResult(new MethodResponse(Encoding.UTF8.GetBytes(DeviceResponseJson), 200));
-                },
-                null).ConfigureAwait(false);
-
-            return methodCallReceived.Task;
-        }
-
-        public static Task<Task> SetDeviceReceiveMethodObsoleteHandler(DeviceClient deviceClient)
-        {
-            var methodCallReceived = new TaskCompletionSource<bool>();
-
-#pragma warning disable CS0618
-            deviceClient.SetMethodHandler(MethodName, (request, context) =>
-            {
-                s_log.WriteLine($"{nameof(SetDeviceReceiveMethodObsoleteHandler)}: DeviceClient method: {request.Name} {request.ResponseTimeout}.");
-
-                try
-                {
-                    Assert.AreEqual(MethodName, request.Name);
-                    Assert.AreEqual(ServiceRequestJson, request.DataAsJson);
-
-                    methodCallReceived.SetResult(true);
-                }
-                catch (Exception ex)
-                {
-                    methodCallReceived.SetException(ex);
-                }
-
-                return Task.FromResult(new MethodResponse(Encoding.UTF8.GetBytes(DeviceResponseJson), 200));
-            }, null);
-#pragma warning restore CS0618
-
-            return Task.FromResult<Task>(methodCallReceived.Task);
-        }
-    }
-}
